storage-capacity: add tests, fix int -> int64_t

+++ src/infinit/storage/Adb.hh
@@ -32,7 +32,7 @@ namespace infinit
     struct AdbStorageConfig
       : public StorageConfig
     {
-      AdbStorageConfig(std::string name, int capacity = 0);
+      AdbStorageConfig(std::string name, int64_t capacity = 0);
       AdbStorageConfig(elle::serialization::SerializerIn& input);
 
       virtual
